Added Global Reference to WeaponController.cs
Added a global reference to the WeaponController.cs class (this needs to be changed if multiple players are added). Added ammo counter to player.

+++ Assets/Scripts/ScriptableWeapons.cs
@@ -24,7 +24,7 @@ public class ScriptableWeapons : ScriptableObject
 
     // In ticks
     public float reloadTime;
-    public float maxAmmo;
+    public int maxAmmo;
 
     // 50 is *really* high
     public float bulletVelocity;
